Added the .bat script
This is just the .bat that simply uses the command I got from Sumit D - VM from Microsoft Answers Community & then restarts Windows Explorer to make the changes take effect instantly. It's very simple & this is mostly just a backup just in case I lose it later on, but anyone is free to use it as they please. -nuckat